migrated existing device inventory to new device inventory.
