+++ cyences_app_for_splunk/default/collections.conf
@@ -26,10 +26,9 @@ accelerated_fields.incident_id = {"notable_event_id": 1}
 
 [cs_all_vuln_collection]
 
+
 # Device Inventory
 [cs_device_inventory_collection]
-
-[cs_device_inventory_temp_collection]
 
 [cs_lansweeper_inventory_collection]
 
@@ -46,26 +45,6 @@ accelerated_fields.incident_id = {"notable_event_id": 1}
 [cs_crowdstrike_inventory_collection]
 
 [cs_kaspersky_inventory_collection]
-
-
-# Device Inventory v2
-[cs_device_inventory_collection_v2]
-
-[cs_lansweeper_inventory_collection_v2]
-
-[cs_tenable_inventory_collection_v2]
-
-[cs_tenable_vuln_collection_v2]
-
-[cs_qualys_inventory_collection_v2]
-
-[cs_sophos_inventory_collection_v2]
-
-[cs_windows_defender_inventory_collection_v2]
-
-[cs_crowdstrike_inventory_collection_v2]
-
-[cs_kaspersky_inventory_collection_v2]
 
 
 # User Inventory
