+++ cyences_app_for_splunk/default/commands.conf
@@ -24,12 +24,6 @@ python.version = python3
 filename = sophos_instance_details.py
 chunked = true
 
-[deviceinventorygen]
-python.version = python3
-filename = device_inventory_gen.py
-is_risky = true
-chunked = true
-
 [cyencesdevicemanager]
 python.version = python3
 filename = cyences_device_manager.py
